+++ ArmPkg/Library/ArmExceptionLib/AArch64/ExceptionSupport.S
@@ -7,7 +7,7 @@
 //
 //------------------------------------------------------------------------------
 
-#include <Chipset/AArch64.h>
+#include <AArch64/AArch64.h>
 #include <Library/PcdLib.h>
 #include <AsmMacroIoLibV8.h>
 #include <Protocol/DebugSupport.h> // for exception type definitions
