+++ ArmPlatformPkg/PrePeiCore/AArch64/Exception.S
@@ -5,7 +5,7 @@
 #
 #
 
-#include <Chipset/AArch64.h>
+#include <AArch64/AArch64.h>
 #include <AsmMacroIoLibV8.h>
 #include <Base.h>
 #include <AutoGen.h>
