+++ ArmVirtPkg/PrePi/AArch64/ArchPrePi.c
@@ -8,7 +8,7 @@
 
 #include "PrePi.h"
 
-#include <Chipset/AArch64.h>
+#include <AArch64/AArch64.h>
 
 VOID
 ArchInitialize (
