+++ MdePkg/Include/Library/ArmLib.h
@@ -16,7 +16,7 @@
 #ifdef MDE_CPU_ARM
   #include <Chipset/ArmV7.h>
 #elif defined (MDE_CPU_AARCH64)
-  #include <Chipset/AArch64.h>
+  #include <AArch64/AArch64.h>
 #else
   #error "Unknown chipset."
 #endif
